fix: fix MatchingDeclarationName warnings
