+++ core/src/com/unciv/models/simulation/MutableInt.kt
